+++ src/main/kotlin/com/github/hotm/world/auranet/AuraNodeUtils.kt
@@ -3,8 +3,8 @@ package com.github.hotm.world.auranet
 import net.minecraft.util.math.BlockPos
 
 object AuraNodeUtils {
-    fun calculateSiphonValue(initDenom: Float, denomStep: Float, chunkAura: Int, siphonCount: Int): Int {
-        return chunkAura / (denomStep * siphonCount + initDenom - denomStep).toInt()
+    fun calculateSiphonValue(initDenom: Float, denomStep: Float, chunkAura: Float, siphonCount: Int): Float {
+        return chunkAura / (denomStep * siphonCount + initDenom - denomStep)
     }
 
     inline fun <reified T> nodeAt(pos: BlockPos?, access: AuraNetAccess): T? {
