+++ src/main/kotlin/com/github/hotm/world/auranet/SourceAuraNode.kt
@@ -1,5 +1,5 @@
 package com.github.hotm.world.auranet
 
 interface SourceAuraNode : AuraNode {
-    fun getSourceAura(): Int
+    fun getSourceAura(): Float
 }
